requirements.txt
opencv-python>=4.8.0
numpy>=1.26.0
pillow>=10.0.0
pytesseract>=0.3.10
albumentations>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.66.0
